Move synbBackupTimeoutPolicy to private API
