+++ hazelcast/src/main/java/com/hazelcast/config/MapConfig.java
@@ -16,7 +16,6 @@
 
 package com.hazelcast.config;
 
-import com.hazelcast.internal.cluster.Versions;
 import com.hazelcast.internal.config.ConfigDataSerializerHook;
 import com.hazelcast.internal.config.DataPersistenceAndHotRestartMerger;
 import com.hazelcast.internal.partition.IPartition;
@@ -24,7 +23,6 @@ import com.hazelcast.map.IMap;
 import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
 import com.hazelcast.nio.serialization.IdentifiedDataSerializable;
-import com.hazelcast.nio.serialization.impl.Versioned;
 
 import javax.annotation.Nonnull;
 import java.io.IOException;
@@ -43,7 +41,7 @@ import static com.hazelcast.internal.util.Preconditions.isNotNull;
 /**
  * Contains the configuration for an {@link IMap}.
  */
-public class MapConfig implements IdentifiedDataSerializable, NamedConfig, Versioned {
+public class MapConfig implements IdentifiedDataSerializable, NamedConfig {
 
     /**
      * The minimum number of backups
@@ -116,59 +114,6 @@ public class MapConfig implements IdentifiedDataSerializable, NamedConfig, Versi
     private boolean perEntryStatsEnabled = DEFAULT_ENTRY_STATS_ENABLED;
     private int backupCount = DEFAULT_BACKUP_COUNT;
     private int asyncBackupCount = MIN_BACKUP_COUNT;
-    private SyncBackupTimeoutPolicy syncBackupTimeoutPolicy = SyncBackupTimeoutPolicy.GLOBAL;
-
-    /**
-     * Specifies action takes when {@link
-     * com.hazelcast.spi.properties.ClusterProperty#OPERATION_BACKUP_TIMEOUT_MILLIS}
-     * elapses when waiting for completion of synchronous backups.
-     */
-    public enum SyncBackupTimeoutPolicy {
-        /**
-         * Use behavior configured by global property:
-         * {@link com.hazelcast.spi.properties.ClusterProperty#FAIL_ON_INDETERMINATE_OPERATION_STATE} or
-         * {@link com.hazelcast.client.properties.ClientProperty#FAIL_ON_INDETERMINATE_OPERATION_STATE}
-         */
-        GLOBAL(0),
-        /**
-         * In case of timeout for sync backup return success
-         */
-        SUCCESS(1),
-        /**
-         * In case of timeout for sync backup throw {@link com.hazelcast.core.IndeterminateOperationStateException}
-         */
-        INDETERMINATE(2);
-
-        private final int id;
-
-        SyncBackupTimeoutPolicy(int id) {
-            this.id = id;
-        }
-
-        /**
-         * Gets the ID for the given {@link SyncBackupTimeoutPolicy}.
-         *
-         * @return the ID
-         */
-        public int getId() {
-            return id;
-        }
-
-        /**
-         * Returns the {@link SyncBackupTimeoutPolicy} as an enum.
-         *
-         * @return the {@link SyncBackupTimeoutPolicy} as an enum
-         */
-        public static SyncBackupTimeoutPolicy getById(final int id) {
-            for (SyncBackupTimeoutPolicy type : values()) {
-                if (type.id == id) {
-                    return type;
-                }
-            }
-            return null;
-        }
-    }
-
     private int timeToLiveSeconds = DEFAULT_TTL_SECONDS;
     private int maxIdleSeconds = DEFAULT_MAX_IDLE_SECONDS;
     private String name;
@@ -207,7 +152,6 @@ public class MapConfig implements IdentifiedDataSerializable, NamedConfig, Versi
         this.name = config.name;
         this.backupCount = config.backupCount;
         this.asyncBackupCount = config.asyncBackupCount;
-        this.syncBackupTimeoutPolicy = config.syncBackupTimeoutPolicy;
         this.timeToLiveSeconds = config.timeToLiveSeconds;
         this.maxIdleSeconds = config.maxIdleSeconds;
         this.metadataPolicy = config.metadataPolicy;
@@ -362,36 +306,6 @@ public class MapConfig implements IdentifiedDataSerializable, NamedConfig, Versi
      */
     public int getTotalBackupCount() {
         return backupCount + asyncBackupCount;
-    }
-
-    /**
-     * Gets action to be taken when {@link
-     * com.hazelcast.spi.properties.ClusterProperty#OPERATION_BACKUP_TIMEOUT_MILLIS}
-     * elapses when waiting for completion of synchronous backups.
-     *
-     * @return action to be taken on backup timeout
-     * @see #setSyncBackupTimeoutPolicy(SyncBackupTimeoutPolicy)
-     */
-    public SyncBackupTimeoutPolicy getSyncBackupTimeoutPolicy() {
-        return syncBackupTimeoutPolicy;
-    }
-
-    /**
-     * Sets action to be taken when {@link
-     * com.hazelcast.spi.properties.ClusterProperty#OPERATION_BACKUP_TIMEOUT_MILLIS}
-     * elapses when waiting for completion of synchronous backups.
-     * <p>
-     * This setting applies only to sync and asynch operations on single key
-     * (eg. {@link IMap#put(Object, Object)}. It does not affect multi-entry
-     * operations (eg. {@link IMap#clear()}, {@link IMap#putAll}).
-     *
-     * @param syncBackupTimeoutPolicy action to be taken on backup timeout
-     * @return the updated MapConfig
-     * @see #setBackupCount(int)
-     */
-    public MapConfig setSyncBackupTimeoutPolicy(SyncBackupTimeoutPolicy syncBackupTimeoutPolicy) {
-        this.syncBackupTimeoutPolicy = syncBackupTimeoutPolicy;
-        return this;
     }
 
     /**
@@ -911,9 +825,6 @@ public class MapConfig implements IdentifiedDataSerializable, NamedConfig, Versi
         if (asyncBackupCount != that.asyncBackupCount) {
             return false;
         }
-        if (syncBackupTimeoutPolicy != that.syncBackupTimeoutPolicy) {
-            return false;
-        }
         if (timeToLiveSeconds != that.timeToLiveSeconds) {
             return false;
         }
@@ -998,7 +909,6 @@ public class MapConfig implements IdentifiedDataSerializable, NamedConfig, Versi
         int result = (name != null ? name.hashCode() : 0);
         result = 31 * result + backupCount;
         result = 31 * result + asyncBackupCount;
-        result = 31 * result + syncBackupTimeoutPolicy.hashCode();
         result = 31 * result + timeToLiveSeconds;
         result = 31 * result + maxIdleSeconds;
         result = 31 * result + evictionConfig.hashCode();
@@ -1035,7 +945,6 @@ public class MapConfig implements IdentifiedDataSerializable, NamedConfig, Versi
                 + ", metadataPolicy=" + metadataPolicy
                 + ", backupCount=" + backupCount
                 + ", asyncBackupCount=" + asyncBackupCount
-                + ", syncBackupTimeoutPolicy=" + syncBackupTimeoutPolicy
                 + ", timeToLiveSeconds=" + timeToLiveSeconds
                 + ", maxIdleSeconds=" + maxIdleSeconds
                 + ", readBackupData=" + readBackupData
@@ -1100,12 +1009,6 @@ public class MapConfig implements IdentifiedDataSerializable, NamedConfig, Versi
         out.writeBoolean(perEntryStatsEnabled);
         out.writeObject(dataPersistenceConfig);
         out.writeObject(tieredStoreConfig);
-
-        // RU_COMPAT 5.2
-        if (out.getVersion().isGreaterOrEqual(Versions.V5_3)) {
-            out.writeInt(syncBackupTimeoutPolicy.getId());
-        }
-
     }
 
     @Override
@@ -1138,10 +1041,5 @@ public class MapConfig implements IdentifiedDataSerializable, NamedConfig, Versi
         perEntryStatsEnabled = in.readBoolean();
         setDataPersistenceConfig(in.readObject());
         setTieredStoreConfig(in.readObject());
-
-        // RU_COMPAT 5.2
-        if (in.getVersion().isGreaterOrEqual(Versions.V5_3)) {
-            syncBackupTimeoutPolicy = SyncBackupTimeoutPolicy.getById(in.readInt());
-        }
     }
 }
